[DM-1159] 'New Opportunity' breadcrumb
When creating a new opportunity, the breadcrumb should read “new
opportunity.”

Now it does, until a user puts anything into the ‘description’ blank,
and that goes in instead.

Uses the incredibly simple ‘mg-hide’ and ‘ng-show’ directives to check
to see if master.discussion is truthy. If so, no-show shows
master.discussion, if not, ng-show shows New Customer.